# spec/openapi.yaml
swagger: '2.0'
info:
  title: AI Gateway
  version: 1.0.0
  description: |
    A platform for machine learning & high performance computing
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http
host: 'localhost:9000'
basePath: /api/v1
securityDefinitions:
  api-authorizer:
    name: Authorization
    type: apiKey
    in: header
paths:
  '/configurations':
    get:
      description: |
        returns app configurations
      operationId: getConfigurations
      tags:
        - app
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Configuration'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      description: |
        set app configurations
      operationId: postConfigurations
      tags:
        - app
      parameters:
        - name: body
          in: body
          required: true
          schema:
            title: AccountInfo
            type: object
            properties:
              docker_registry:
                description: Docker Registry endpoint
                type: string
              docker_hostname:
                description: Hostname for the private Docker registry
                type: string
              docker_username:
                description: Username for the private Docker registry
                type: string
              docker_password:
                description: Password for the private Docker registry
                type: string
                format: password
              ngc_email:
                description: E-mail address for NGC console
                type: string
                format: email
              ngc_password:
                description: Password for NGC console
                type: string
                format: password
              ngc_apikey:
                description: NGC - API Key
                type: string
              k8s_config:
                description: kubecfg
                type: string
              rescale_platform:
                description: Rescale platform endopoint
                type: string
                enum:
                  - https://platform.rescale.com
                  - https://platform.rescale.jp
                  - https://kr.rescale.com
                  - https://eu.rescale.com
              rescale_key:
                description: Rescale - API Key
                type: string
      responses:
        '201':
          description: OK
          schema:
            $ref: '#/definitions/Session'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/sessions':
    post:
      description: |
        login
      operationId: postNewSession
      tags:
        - app
      parameters:
        - name: body
          in: body
          required: true
          schema:
            title: AccountInfo
            type: object
            required:
              - docker_username
              - docker_password
            properties:
              docker_username:
                description: Username for the private Docker registry
                type: string
              docker_password:
                description: Password for the private Docker registry
                type: string
                format: password
      responses:
        '201':
          description: OK
          schema:
            $ref: '#/definitions/Session'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  /jobs:
    get:
      description: |
        returns training jobs on cloud
      operationId: getJobs
      tags:
        - job
      security:
        - api-authorizer: []
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Job'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      description: |
        Submit a job with the specified image
      operationId: postNewJob
      tags:
        - job
      security:
        - api-authorizer: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            title: JobAttrs
            type: object
            properties:
              platform_id:
                description: Platform ID
                type: string
                enum:
                    - kubernetes
                    - rescale
              notebook_id:
                description: Notebook container ID
                type: string
              entrypoint_file:
                description: The entrypoint file of the job
                type: string
              commands:
                description: commands to be excuted after the entrypoint
                type: array
                items:
                  type: string
              cpu:
                description: Requesting millicores of CPU
                type: integer
                format: int64
              mem:
                description: Requesting bytes of memory
                type: integer
                format: int64
              gpu:
                description: Requesting number of GPU
                type: integer
                format: int64
              coretype:
                description: Rescale CoreType as its infrastructure
                type: string
              cores:
                description: The number of CPU cores
                type: integer
                format: int64
      responses:
        '201':
          description: OK
          schema:
            title: newJobID
            type: object
            properties:
              id:
                description: ID of the new job
                type: string
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  '/workspaces':
    get:
      description: |
        returns user's workspaces
      operationId: getWorkspaces
      tags:
        - workspace
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Workspace'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      description: |
        delete user's workspace
      operationId: deleteWorkspace
      tags:
        - workspace
      parameters:
        - name: body
          in: body
          required: true
          schema:
            title: workspace
            type: object
            required:
              - path
            properties:
              path:
                type: string
      responses:
        '204':
          description: OK
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
  Configuration:
    description: app's configurations
    required:
      - must_signed_in
    properties:
      must_signed_in:
        description: Users should be signed in
        type: string
        enum:
          - "yes"
          - "no"
      use_private_registry:
        description: Private registry will be used or not
        type: string
        enum:
          - "yes"
          - "no"
      use_ngc:
        description: NGC will be used or not
        type: string
        enum:
          - "yes"
          - "no"
      use_k8s:
        description: Kubernetes will be used or not
        type: string
        enum:
          - "yes"
          - "no"
      use_rescale:
        description: Rescale will be used or not
        type: string
        enum:
          - "yes"
          - "no"
      docker_registry:
        description: Docker Registry endpoint
        type: string
      docker_hostname:
        description: Hostname for the private Docker registry
        type: string
      docker_username:
        description: Username for the private Docker registry
        type: string
      docker_password:
        description: Fist 3 chars of the password of the private Docker registry
        type: string
      ngc_email:
        description: E-mail address for NGC console
        type: string
        format: email
      ngc_password:
        description: Fist 3 chars of the password for NGC console
        type: string
      ngc_apikey:
        description: Fist 5 chars of NGC API Key
        type: string
      k8s_config:
        description: kubecfg
        type: string
      rescale_platform:
        type: string
        enum:
          - https://platform.rescale.com
          - https://platform.rescale.jp
          - https://kr.rescale.com
          - https://eu.rescale.com
      rescale_key:
        description: Fist 5 chars of Rescal API Key
        type: string
  Session:
    description: docker image information
    required:
      - token
    properties:
      token:
        description: API Token
        type: string
...
  Job:
    description: Rescale Job information
    required:
      - id
    properties:
      id:
        description: Job ID
        type: string
      platform:
        description: platform
        type: string
        enum:
          - kubernetes
          - rescale
      status:
        description: the status of the job
        type: string
      image:
        description: the image ID
        type: string
      mounts:
        description: the container labels
        type: array
        items:
          type: string
      commands:
        description: the container labels
        type: array
        items:
          type: string
      started:
        description: started unix timestamp
        type: string
        format: date-time
      ended:
        description: ended unix timestamp
        type: string
        format: date-time
      external_link:
        description: A link to an external status page
        type: string
...
  Workspace:
    description: User's workspace
    required:
      - path
    properties:
      notebooks:
        description: Notebooks which are mounting the workspace
        type: array
        items:
          type: string
      jobs:
        description: Jobs which are mounting the workspace
        type: array
        items:
          type: string
      path:
        type: string
      absolute_path:
        type: string
...
  Error:
    required:
      - code
      - message
    properties:
      code:
        type: string
      message:
        type: string